feat(SearchPage): component split
