+++ packages/mirror-tv-next/app/search/[keyword]/page.tsx
@@ -1,6 +1,4 @@
 import React from 'react'
-import styles from '~/styles/pages/search-page.module.scss'
-import UiPostCard from '~/components/shared/ui-post-card'
 import SearchNoResult from '~/components/search/search-no-result'
 import { POPULAR_POSTS_URL } from '~/constants/environment-variables'
 import {
@@ -9,6 +7,7 @@ import {
   TVPost,
   TVPostResponse,
 } from '~/types/api-data'
+import SearchResult from '~/components/search/search-result'
 // write typescript type of params
 type slug = {
   params: params
@@ -35,7 +34,6 @@ const getPopularResult = async (): Promise<PopularSearchItemResponse> => {
 
 const page = async ({ params }: slug) => {
   const keyword = decodeURI(params.keyword)
-  const searchNoResultProps = { width: 166, height: 204 }
   let searchResultList: TVPost[] = []
   let popularResultList: PopularSearchItem[] = []
   try {
@@ -52,70 +50,20 @@ const page = async ({ params }: slug) => {
     console.error('searchResultList_error', e)
   }
 
+  // update after fetch
+  const searchNoResultProps = {
+    width: 166,
+    height: 204,
+    keyword,
+    popularResultList,
+  }
+  const searchResultProps = {
+    keyword,
+    searchResultList,
+  }
   if (searchResultList.length === 0)
-    return (
-      <main className={styles.main}>
-        <SearchNoResult {...searchNoResultProps} />
-        <section className={styles.noResultTxt}>
-          <p>您搜尋的「{keyword}」</p>
-          <p>查無相關結果</p>
-        </section>
-        <div className={styles.borderdHeadingWrapper}>
-          <p>熱門新聞</p>
-        </div>
-        <div className={styles.divider} />
-        <ul className={styles.popularResultList}>
-          {popularResultList.map((popularPost) => {
-            const date = new Date(popularPost.publishTime)
-            const props = {
-              title: popularPost.name,
-              date,
-              href: `/story/${popularPost.slug}`,
-              images: {
-                original: popularPost.heroImage?.urlMobileSized || '',
-                w400: popularPost.heroImage?.urlMobileSized || '',
-              },
-              postStyle: 'article',
-              mobileLayoutDirection: 'column' as 'column' | 'row',
-              postTitleHighlightText: '',
-            }
-            return (
-              <li key={popularPost.id}>
-                <UiPostCard {...props} />
-              </li>
-            )
-          })}
-        </ul>
-      </main>
-    )
-  return (
-    <main className={styles.main}>
-      <p className={styles.searchKeyword}>{keyword}</p>
-      <ul className={styles.searchResultList}>
-        {searchResultList.map((searchResult) => {
-          const date = new Date(searchResult._source.publishTime)
-          const props = {
-            title: searchResult._source.name,
-            date,
-            href: `/story/${searchResult._source.slug}`,
-            images: {
-              original: searchResult._source.heroImage.urlMobileSized,
-              w400: searchResult._source.heroImage.urlMobileSized,
-            },
-            postStyle: 'article',
-            mobileLayoutDirection: 'column' as 'column' | 'row',
-            postTitleHighlightText: '',
-          }
-          return (
-            <li key={searchResult._id}>
-              <UiPostCard {...props} />
-            </li>
-          )
-        })}
-        <button className={styles.seeMoreBtn}>看更多</button>
-      </ul>
-    </main>
-  )
+    return <SearchNoResult {...searchNoResultProps} />
+  return <SearchResult {...searchResultProps} />
 }
 
 export default page
